提交于9月8日 20:10 基本功能完善 「登录，注册。。。。」

+++ carmanager-mapper/src/main/java/dqyy/AccountInfoMapper.java
@@ -11,7 +11,7 @@ public interface AccountInfoMapper {
 
     int insertSelective(AccountInfo record);
 
-    AccountInfo selectByPrimaryKey(Byte id);
+    AccountInfo selectByPrimaryKey(Integer id);
 
     List<AccountInfo> selectAll();
